Moving GraknEngineConfig class
