+++ grakn-engine/src/main/java/ai/grakn/engine/controller/CommitLogController.java
@@ -20,7 +20,7 @@ package ai.grakn.engine.controller;
 
 import ai.grakn.concept.ConceptId;
 import ai.grakn.engine.postprocessing.EngineCache;
-import ai.grakn.engine.util.GraknEngineConfig;
+import ai.grakn.engine.GraknEngineConfig;
 import ai.grakn.exception.GraknEngineServerException;
 import ai.grakn.util.ErrorMessage;
 import ai.grakn.util.REST;
