+++ grakn-engine/src/main/java/ai/grakn/engine/controller/StatusController.java
@@ -18,9 +18,8 @@
 
 package ai.grakn.engine.controller;
 
-import ai.grakn.engine.util.GraknEngineConfig;
+import ai.grakn.engine.GraknEngineConfig;
 import ai.grakn.util.REST;
-import com.google.common.collect.Maps;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import mjson.Json;
