+++ grakn-engine/src/main/java/ai/grakn/engine/loader/LoaderTask.java
@@ -21,7 +21,7 @@ package ai.grakn.engine.loader;
 import ai.grakn.GraknGraph;
 import ai.grakn.engine.tasks.BackgroundTask;
 import ai.grakn.engine.postprocessing.EngineCache;
-import ai.grakn.engine.util.GraknEngineConfig;
+import ai.grakn.engine.GraknEngineConfig;
 import ai.grakn.exception.GraknValidationException;
 import ai.grakn.factory.EngineGraknGraphFactory;
 import ai.grakn.graql.Graql;
@@ -37,7 +37,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.function.Consumer;
 
-import static ai.grakn.engine.util.GraknEngineConfig.LOADER_REPEAT_COMMITS;
+import static ai.grakn.engine.GraknEngineConfig.LOADER_REPEAT_COMMITS;
 import static ai.grakn.util.ErrorMessage.FAILED_VALIDATION;
 import static ai.grakn.util.ErrorMessage.ILLEGAL_ARGUMENT_EXCEPTION;
 import static ai.grakn.util.REST.Request.KEYSPACE_PARAM;
