+++ grakn-engine/src/main/java/ai/grakn/engine/tasks/manager/StandaloneTaskManager.java
@@ -26,7 +26,7 @@ import ai.grakn.engine.tasks.TaskSchedule;
 import ai.grakn.engine.tasks.TaskState;
 import ai.grakn.engine.tasks.TaskStateStorage;
 import ai.grakn.engine.tasks.storage.TaskStateInMemoryStore;
-import ai.grakn.engine.util.GraknEngineConfig;
+import ai.grakn.engine.GraknEngineConfig;
 import ai.grakn.engine.util.EngineID;
 import javafx.util.Pair;
 import org.slf4j.Logger;
