+++ grakn-engine/src/main/java/ai/grakn/engine/tasks/manager/ZookeeperConnection.java
@@ -20,7 +20,7 @@
 package ai.grakn.engine.tasks.manager;
 
 import ai.grakn.engine.tasks.config.ZookeeperPaths;
-import ai.grakn.engine.util.GraknEngineConfig;
+import ai.grakn.engine.GraknEngineConfig;
 import ai.grakn.exception.EngineStorageException;
 import org.apache.curator.framework.CuratorFramework;
 import org.apache.curator.framework.CuratorFrameworkFactory;
@@ -35,11 +35,11 @@ import static ai.grakn.engine.tasks.config.ZookeeperPaths.FAILOVER;
 import static ai.grakn.engine.tasks.config.ZookeeperPaths.SCHEDULER;
 import static ai.grakn.engine.tasks.config.ZookeeperPaths.TASKS_PATH_PREFIX;
 import static ai.grakn.engine.tasks.config.ZookeeperPaths.TASK_LOCK_SUFFIX;
-import static ai.grakn.engine.util.GraknEngineConfig.ZK_BACKOFF_BASE_SLEEP_TIME;
-import static ai.grakn.engine.util.GraknEngineConfig.ZK_BACKOFF_MAX_RETRIES;
-import static ai.grakn.engine.util.GraknEngineConfig.ZK_CONNECTION_TIMEOUT;
-import static ai.grakn.engine.util.GraknEngineConfig.ZK_SERVERS;
-import static ai.grakn.engine.util.GraknEngineConfig.ZK_SESSION_TIMEOUT;
+import static ai.grakn.engine.GraknEngineConfig.ZK_BACKOFF_BASE_SLEEP_TIME;
+import static ai.grakn.engine.GraknEngineConfig.ZK_BACKOFF_MAX_RETRIES;
+import static ai.grakn.engine.GraknEngineConfig.ZK_CONNECTION_TIMEOUT;
+import static ai.grakn.engine.GraknEngineConfig.ZK_SERVERS;
+import static ai.grakn.engine.GraknEngineConfig.ZK_SESSION_TIMEOUT;
 
 /**
  * <p>
