+++ grakn-engine/src/main/java/ai/grakn/engine/util/JWTHandler.java
@@ -18,6 +18,7 @@
 
 package ai.grakn.engine.util;
 
+import ai.grakn.engine.*;
 import ai.grakn.exception.GraknEngineServerException;
 import com.auth0.jwt.JWTSigner;
 import com.auth0.jwt.JWTVerifier;
