+++ grakn-test/src/test/java/ai/grakn/test/engine/GraknEngineServerTest.java
@@ -24,15 +24,15 @@ import ai.grakn.engine.tasks.manager.multiqueue.MultiQueueTaskManager;
 import ai.grakn.engine.tasks.manager.singlequeue.SingleQueueTaskManager;
 import ai.grakn.engine.tasks.storage.TaskStateGraphStore;
 import ai.grakn.engine.tasks.storage.TaskStateZookeeperStore;
-import ai.grakn.engine.util.GraknEngineConfig;
+import ai.grakn.engine.GraknEngineConfig;
 import ai.grakn.test.EngineContext;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
-import static ai.grakn.engine.util.GraknEngineConfig.TASK_MANAGER_IMPLEMENTATION;
-import static ai.grakn.engine.util.GraknEngineConfig.USE_ZOOKEEPER_STORAGE;
-import static ai.grakn.engine.util.GraknEngineConfig.ZK_CONNECTION_TIMEOUT;
+import static ai.grakn.engine.GraknEngineConfig.TASK_MANAGER_IMPLEMENTATION;
+import static ai.grakn.engine.GraknEngineConfig.USE_ZOOKEEPER_STORAGE;
+import static ai.grakn.engine.GraknEngineConfig.ZK_CONNECTION_TIMEOUT;
 import static ai.grakn.test.GraknTestEnv.ensureCassandraRunning;
 import static junit.framework.TestCase.assertTrue;
 import static org.hamcrest.MatcherAssert.assertThat;
